SniffingTool/requirements.txt
streamlit>=1.27.0
PyPDF2>=3.0.1
google-generativeai>=0.3.2
pandas>=2.0.0
pdfplumber>=0.10.2
